# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Doctor API",
        "description": "API for managing doctors",
        "contact": {
            "email": "murtada199815@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/AddPatient": {
            "post": {
                "tags": [
                    "Patient"
                ],
                "summary": "add Patient",
                "description": "Register Patient ",
                "operationId": "add-Patient",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Set user authentication token",
                        "schema": {
                            "type": "beraer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "doctor",
                                    "specialization",
                                    "date",
                                    "email",
                                    "phone",
                                    "patientName",
                                    "address"
                                ],
                                "properties": {
                                    "doctor": {
                                        "type": "string"
                                    },
                                    "specialization": {
                                        "type": "string"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "email": {
                                        "type": "email"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "patientName": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "patient created successfully",
                        "content": {
                            "multipart/form-data": {
                                "schema": {
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer"
                                        },
                                        "patient": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "No Data"
                    }
                }
            }
        },
        "/api/GetPatient": {
            "get": {
                "tags": [
                    "Patient"
                ],
                "summary": "get Patient",
                "description": "get Patient ",
                "operationId": "get-Patient",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Set user authentication token",
                        "schema": {
                            "type": "beraer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "": {
                                        "type": ""
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successfully",
                        "content": {
                            "multipart/form-data": {
                                "schema": {
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer"
                                        },
                                        "patient": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Data Not Found!"
                    }
                }
            }
        },
        "/api/ActiveStatusStatus/{id}": {
            "put": {
                "tags": [
                    "Patient"
                ],
                "summary": "update Patient",
                "description": "update  Patient  Here",
                "operationId": "ActiveStatusPatient",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Set user authentication token",
                        "schema": {
                            "type": "beraer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "specialization": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Patient updated successfully",
                        "content": {
                            "multipart/form-data": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Data Not Found!"
                    }
                }
            }
        },
        "/api/InActiveStatusStatus/{id}": {
            "put": {
                "tags": [
                    "Patient"
                ],
                "summary": "update Patient",
                "description": "update  Patient  Here",
                "operationId": "InActiveStatusPatient",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Set user authentication token",
                        "schema": {
                            "type": "beraer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "specialization": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Patient updated successfully",
                        "content": {
                            "multipart/form-data": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Data Not Found!"
                    }
                }
            }
        },
...
    },
...
}